combined aia and hmi image classes

+++ sdo_pypline/sdo_io.py
@@ -7,5 +7,5 @@ def read_header(file):
     return fits.getheader(file, 1)
 
 def read_data(file):
-    return fits.getdata(file, 1)
+    return fits.getdata(file, 1).astype(float)
